+++ CHANGELOG.md
@@ -12,6 +12,7 @@ major job run versions.
 ## [Unreleased]
 ### Added
 - WorkHandler for depth-first listing.
+- Internal-testing flag.
 
 ## [0.0.0] - 2018-12-12
 ### Added